Deleted BundlesScreen.kt to make PR work - should b more readable or something
